solver: make solving in msdf faster
For some strange reason it is a lot faster to reverse the automata,
apply minimization, and then reverse it back if using
most-significant-bit-first representation. Since we are using it let us
employ this strange observation to make kind of minimization really
faster. It will no longer be real minimization in the sense the
resulting automaton will not be deterministic but it may become really
smaller than the original one.

Closes #63