Fix a bug in filtering orderings & do some cleanup
For some reason, `2**x + x = 5` has been broken. After introducing in
commit 9f8de8b ('switch to MSB from LSB') ordering filter ensuring there
is at least one path when the ordering is applied projected all of the
variables making the expression satisfiable. The latter made such
expressions as `2**x + x = 5` sat. Let us disable filtering logic to fix
it.

Also, this was supposed to be a cleanup patch right before merging
smtml migration to main. But turns out there is a few problems. It adds
a lot of new LIA and QF_EIA tests and fixes printing labels.