nfa: speedup MSB-nat invariants update
For MSB NFAs to work properly starting states should be extended after
some kinds of operations over them. Previous implementation of the
extension function was too straitforward: it used BFS to traverse the
NFA using lists inside. The latter fact made the prodedure traverse the
same states many times in case of the dense NFA. Let's use sets in that
BFS procedure to speedup the calculations and not repeat them over and
over.