solver: distinguish integers and strings in models
This patch introduces logic that distinguishes strings and integers in
the model trying to infer the type from `declare-const` and
`declare-funs`.

Also, it filters internal variables from the model. Probably it is
rather dirty since internal variables are just plain variables that
start from the precent sign. This should be fixed in the further
patches.