solver: unify integer and string in string mode
This patch is an attempt to unify the strings and the integer
representations while working in the QF_S (i.e. string) solver mode. The
idea is to build the automaton for integer `=` and `<=` including `$`
(i.e. end of string) in them and to update final states w.r.t. `$`
labels.

Unfortunately, writing longer description are too non-trivial. So,
please, check the code for clarification.