overapprox: do not run with forall quantifiers
This patch makes overapproximation do not run for the forumlas involving
`forall` quantifier. Basically the problem is that when sent to
over-approximation procedure such formulas are simplified by Z3. At the
same time Z3 stucks when trying to simplify these `forall` quantifiers.

Let's detect `forall` by seeking `exists -> ... -> not -> ... -> exists`
pattern in AST and do not execute Z3 (and overall over-approximation) in
that case.