Fix bug when simplifier forgots about divisibility
IN EXP-solver benchmarks (for example, test27) we were losing
information. We were supposed to substitute variables if they are not in
exponents, and forget about equation. But after introduction of strings
we need to check that they are neither in exponents, nor in string
expression.

The check is added. EXP-solver tests are timeouting for now.

Signed-off-by: Kakadu <Kakadu@pm.me>