solver: fix handling exponential-only variables
This patch fixes handling of the variables that occur only under the
exponents. They've been broken when used in the most-significant-bit
mode.

The variables that occur only under the power functional symbols can be
projected w.r.t. the fact the resulting valued should satisfy regex
`0*1`. But due to the typo this variables have been left as is and not
handled anyhow ending up with non-trivial automata that can't be handled
by the most-significant-bit-first mode. Let's fix the typo and finally
get rid of such trivial exponentiation.

Corresponding tests will be added later since currently known
problematic tests are not working because of other bugs.